+++ packages/core/src/dom/media/proxy.ts
@@ -1,0 +1,7 @@
+import { ProxyMixin } from '../../core/media/proxy';
+
+export const MediaProxyMixin = ProxyMixin(
+  globalThis.HTMLVideoElement ?? class {},
+  globalThis.HTMLMediaElement ?? class {},
+  globalThis.EventTarget ?? class {}
+);